fix: 커스텀 훅 적용 및 lazy 페이지로 변경
